test(claude-sdk-oauth): resolve ninth cubic review round on probes and spikes
- reattach spike: the secondary account's access token joins the redaction
  list — a cross-account SDK error string could otherwise echo it into the
  REJECTED line
- spike support: the guarded setup + signal-aware cleanup shared by both live
  spikes is extracted into startGuardedQuery (handlers installed BEFORE setup
  so a signal during SDK import/binary resolution is covered; disarm()
  removes them at arm end); both spikes adopt it, deleting their divergent
  copies
- fullstack probe: the loopback handler route-checks POSTs (/messages only,
  404 otherwise) and shape-checks the parsed body (model string + messages
  array, 400 otherwise) so a wrong-route or wrong-shape request can never be
  answered with the success fixture
- registry probe: safeDetail is imported from the fullstack support module
  instead of keeping a second copy
